done build base

+++ src/main/java/com/app/controller/AccountController.java
@@ -6,7 +6,6 @@ import com.app.constant.ErrorCode;
 import com.app.dto.ApiMessageDto;
 import com.app.dto.ResponseListDto;
 import com.app.dto.account.AccountAdminDto;
-import com.app.dto.account.AccountDto;
 import com.app.dto.account.AccountForgetPasswordDto;
 import com.app.form.account.*;
 import com.app.mapper.AccountMapper;
@@ -28,7 +27,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -36,11 +34,11 @@ import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 
+@Slf4j
 @RestController
 @RequestMapping("/v1/account")
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 @FieldDefaults(level = AccessLevel.PRIVATE)
-@Slf4j
 public class AccountController extends ABasicController {
     @Autowired
     AccountRepository accountRepository;
@@ -75,15 +73,16 @@ public class AccountController extends ABasicController {
     }
 
     @PostMapping(value = "/create-admin", produces = MediaType.APPLICATION_JSON_VALUE)
-    @PreAuthorize("hasRole('ACC_C_AD')")
-    public ApiMessageDto<String> createAdmin(@Valid @RequestBody CreateAccountAdminForm createAccountAdminForm, BindingResult bindingResult) {
-        Account accountByUsername = accountRepository.findAccountByUsername(createAccountAdminForm.getUsername()).orElse(null);
-        if (accountByUsername != null) {
+    @PreAuthorize("hasRole('ACC_C_A')")
+    public ApiMessageDto<String> createAdmin(@Valid @RequestBody CreateAccountAdminForm createAccountAdminForm) {
+        if (accountRepository.findFirstByUsername(createAccountAdminForm.getUsername()).isPresent()) {
             return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_USERNAME_EXISTED, "Username existed");
         }
-        Account accountByEmail = accountRepository.findAccountByEmail(createAccountAdminForm.getEmail()).orElse(null);
-        if (accountByEmail != null) {
+        if (accountRepository.findFirstByEmail(createAccountAdminForm.getEmail()).isPresent()) {
             return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_EMAIL_EXISTED, "Email existed");
+        }
+        if (accountRepository.findFirstByPhone(createAccountAdminForm.getPhone()).isPresent()) {
+            return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_PHONE_EXISTED, "Phone existed");
         }
         Group group = groupRepository.findById(createAccountAdminForm.getGroupId()).orElse(null);
         if (group == null) {
@@ -98,29 +97,27 @@ public class AccountController extends ABasicController {
     }
 
     @PutMapping(value = "/update-admin", produces = MediaType.APPLICATION_JSON_VALUE)
-    @PreAuthorize("hasRole('ACC_U_AD')")
-    public ApiMessageDto<String> updateAdmin(@Valid @RequestBody UpdateAccountAdminForm updateAccountAdminForm, BindingResult bindingResult) {
+    @PreAuthorize("hasRole('ACC_U_A')")
+    public ApiMessageDto<String> updateAdmin(@Valid @RequestBody UpdateAccountAdminForm updateAccountAdminForm) {
         Account account = accountRepository.findById(updateAccountAdminForm.getId()).orElse(null);
         if (account == null) {
             return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_NOT_FOUND, "Not found account");
         }
-        if (updateAccountAdminForm.getEmail() != null && !updateAccountAdminForm.getEmail().equals(account.getEmail())) {
-            Account accountByEmail = accountRepository.findAccountByEmail(updateAccountAdminForm.getEmail()).orElse(null);
-            if (accountByEmail != null) {
-                return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_EMAIL_EXISTED, "Email existed");
-            }
+        if (updateAccountAdminForm.getEmail() != null && !updateAccountAdminForm.getEmail().equals(account.getEmail())
+                && accountRepository.findFirstByEmail(updateAccountAdminForm.getEmail()).isPresent()) {
+            return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_EMAIL_EXISTED, "Email existed");
+        }
+        if (updateAccountAdminForm.getPhone() != null && !updateAccountAdminForm.getPhone().equals(account.getPhone())
+                && accountRepository.findFirstByPhone(updateAccountAdminForm.getPhone()).isPresent()) {
+            return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_PHONE_EXISTED, "Phone existed");
         }
         Group group = groupRepository.findById(updateAccountAdminForm.getGroupId()).orElse(null);
         if (group == null) {
             return makeErrorResponse(ErrorCode.GROUP_ERROR_NOT_FOUND, "Not found group");
         }
-        if (StringUtils.isNoneBlank(updateAccountAdminForm.getPassword())) {
-            account.setPassword(passwordEncoder.encode(updateAccountAdminForm.getPassword()));
-        }
-        if (StringUtils.isNoneBlank(updateAccountAdminForm.getAvatar())) {
-            if (!updateAccountAdminForm.getAvatar().equals(account.getAvatar())) {
-                apiService.deleteFile(account.getAvatar());
-            }
+        if (StringUtils.isNotBlank(updateAccountAdminForm.getAvatar())
+                && !updateAccountAdminForm.getAvatar().equals(account.getAvatar())) {
+            apiService.deleteFile(account.getAvatar());
             account.setAvatar(updateAccountAdminForm.getAvatar());
         }
         accountMapper.fromUpdateAccountAdminFormToEntity(updateAccountAdminForm, account);
@@ -137,7 +134,10 @@ public class AccountController extends ABasicController {
             return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_NOT_FOUND, "Not found account");
         }
         if (account.getIsSuperAdmin()) {
-            return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_NOT_ALLOW_DELETE_SUPPER_ADMIN, "Not allow to delete super admin");
+            return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_NOT_ALLOW_DELETE_SUPPER_ADMIN, "Not allowed to delete super admin");
+        }
+        if (Long.valueOf(getCurrentUser()).equals(account.getId())) {
+            return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_NOT_ALLOW_DELETE_YOURSELF, "Not allowed to delete yourself");
         }
         apiService.deleteFile(account.getAvatar());
         accountRepository.deleteById(id);
@@ -145,42 +145,53 @@ public class AccountController extends ABasicController {
     }
 
     @GetMapping(value = "/profile", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ApiMessageDto<AccountDto> profile() {
+    public ApiMessageDto<AccountAdminDto> profile() {
         Account account = accountRepository.findById(getCurrentUser()).orElse(null);
         if (account == null) {
             return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_NOT_FOUND, "Not found account");
         }
-        return makeSuccessResponse(accountMapper.fromEntityToAccountDto(account), "Get profile success");
+        return makeSuccessResponse(accountMapper.fromEntityToAccountAdminDto(account), "Get profile success");
     }
 
     @PutMapping(value = "/update-profile-admin", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ApiMessageDto<String> updateProfileAdmin(@Valid @RequestBody UpdateProfileAdminForm updateProfileAdminForm, BindingResult bindingResult) {
+    public ApiMessageDto<String> updateProfileAdmin(@Valid @RequestBody UpdateProfileAdminForm updateProfileAdminForm) {
         Account account = accountRepository.findById(getCurrentUser()).orElse(null);
         if (account == null) {
             return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_NOT_FOUND, "Not found account");
         }
         if (!passwordEncoder.matches(updateProfileAdminForm.getOldPassword(), account.getPassword())) {
-            return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_WRONG_PASSWORD, "Old password is incorrect");
+            return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_PASSWORD_INVALID, "Old password is incorrect");
         }
-        if (StringUtils.isNoneBlank(updateProfileAdminForm.getPassword())) {
-            account.setPassword(passwordEncoder.encode(updateProfileAdminForm.getPassword()));
-        }
-        if (StringUtils.isNoneBlank(updateProfileAdminForm.getAvatar())) {
-            if (!updateProfileAdminForm.getAvatar().equals(account.getAvatar())) {
-                //delete old image
-                apiService.deleteFile(account.getAvatar());
-            }
-            account.setAvatar(updateProfileAdminForm.getAvatar());
+        String newAvatar = updateProfileAdminForm.getAvatar();
+        if (StringUtils.isNotBlank(newAvatar) && !newAvatar.equals(account.getAvatar())) {
+            apiService.deleteFile(account.getAvatar());
+            account.setAvatar(newAvatar);
         }
         accountMapper.fromUpdateProfileAdminFormToEntity(updateProfileAdminForm, account);
         accountRepository.save(account);
         return makeSuccessResponse(null, "Update profile success");
+    }
 
+    @PutMapping(value = "/change-profile-password", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ApiMessageDto<String> changeProfilePassword(@Valid @RequestBody ChangeProfilePasswordForm changeProfilePasswordForm) {
+        Account account = accountRepository.findById(getCurrentUser()).orElse(null);
+        if (account == null) {
+            return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_NOT_FOUND, "Not found account");
+        }
+        if (!passwordEncoder.matches(changeProfilePasswordForm.getOldPassword(), account.getPassword())) {
+            return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_PASSWORD_INVALID, "Old password is incorrect");
+        }
+        if (changeProfilePasswordForm.getNewPassword().equals(changeProfilePasswordForm.getOldPassword())) {
+            return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_PASSWORD_INVALID, "New password must be different from old password");
+        }
+        account.setPassword(passwordEncoder.encode(changeProfilePasswordForm.getNewPassword()));
+        accountRepository.save(account);
+        return makeSuccessResponse(null, "Change profile password success");
     }
 
     @PostMapping(value = "/request-forget-password", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ApiMessageDto<AccountForgetPasswordDto> requestForgetPassword(@Valid @RequestBody RequestForgetPasswordForm forgetForm, BindingResult bindingResult) {
-        Account account = accountRepository.findAccountByEmail(forgetForm.getEmail()).orElse(null);
+    public ApiMessageDto<AccountForgetPasswordDto> requestForgetPassword(@Valid @RequestBody RequestForgetPasswordForm forgetForm) {
+        Account account = accountRepository.findFirstByEmail(forgetForm.getEmail()).orElse(null);
         if (account == null) {
             return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_NOT_FOUND, "Not found account");
         }
@@ -189,15 +200,15 @@ public class AccountController extends ABasicController {
         account.setResetPasswordCode(otp);
         account.setResetPasswordTime(new Date());
         accountRepository.save(account);
-        apiService.sendEmail(account.getEmail(), "OTP: " + otp, "Request forget password successful, please check email", false);
-        AccountForgetPasswordDto accountForgetPasswordDto = new AccountForgetPasswordDto();
+        apiService.sendEmail(account.getEmail(), "OTP: " + otp, "Request forget password successfully, please check email", false);
         String zipUserId = ZipUtils.zipString(account.getId() + ";" + otp);
+        AccountForgetPasswordDto accountForgetPasswordDto = new AccountForgetPasswordDto();
         accountForgetPasswordDto.setUserId(zipUserId);
-        return makeSuccessResponse(accountForgetPasswordDto, "Request forget password successful, please check email");
+        return makeSuccessResponse(accountForgetPasswordDto, "Request forget password successfully, please check email");
     }
 
     @PostMapping(value = "/reset-password", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ApiMessageDto<String> forgetPassword(@Valid @RequestBody ResetPasswordForm resetForm, BindingResult bindingResult) {
+    public ApiMessageDto<String> forgetPassword(@Valid @RequestBody ResetPasswordForm resetForm) {
         String[] unzip = Objects.requireNonNull(ZipUtils.unzipString(resetForm.getUserId())).split(";", 2);
         Long id = ConvertUtils.convertStringToLong(unzip[0]);
         Account account = accountRepository.findById(id).orElse(null);
@@ -207,8 +218,9 @@ public class AccountController extends ABasicController {
         if (account.getAttemptCode() >= AppConstant.MAX_ATTEMPT_FORGET_PASSWORD) {
             return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_EXCEEDED_NUMBER_OF_INPUT_ATTEMPT_OTP, "Exceeded number of input attempt OTP");
         }
-        if (!account.getResetPasswordCode().equals(resetForm.getOtp()) ||
-                (new Date().getTime() - account.getResetPasswordTime().getTime() >= AppConstant.MAX_ATTEMPT_FORGET_PASSWORD)) {
+        boolean isOtpInvalid = !account.getResetPasswordCode().equals(resetForm.getOtp())
+                || (new Date().getTime() - account.getResetPasswordTime().getTime() >= AppConstant.MAX_ATTEMPT_FORGET_PASSWORD);
+        if (isOtpInvalid) {
             account.setAttemptCode(account.getAttemptCode() + 1);
             accountRepository.save(account);
             return makeErrorResponse(ErrorCode.ACCOUNT_ERROR_OTP_INVALID, "OTP code invalid or has expired");
